pyproject.toml
[project]
name = "fastapi-zero"
version = "0.1.0"
description = ""
authors = [
    {name = "Romario lima",email = "romariolima1006@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.10"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"

[tool.ruff]
line-length = 79
extend-exclude = [
    "migrations",
]

[tool.ruff.lint]
preview = true
select = [
    "F",
    "I",
    "W",
    "E",
    "PL",
    "PT",
    "FAST"
]

[tool.ruff.format]
preview = true
quote-style = "single"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[tool.taskipy.tasks]
lint = "ruff check"
pre_format = "ruff check --fix"
format = "ruff format"
run = "fastapi dev fastapi_zero/app.py"
test = "pytest -s -x --cov=fastapi_zero -vv"
post_test = "coverage html"